apps/workers/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
praw==7.8.1
openai==1.12.0
supabase==2.10.0
python-dotenv==1.0.1
httpx==0.27.0
websockets>=13.0
